requirements.txt

watchdog
pyyaml
